+++ spec/resources/stub_zone_spec.rb
@@ -7,6 +7,8 @@ describe 'adding stub zones' do
     ).converge('bind_test::spec_stub_zone')
   end
 
+  include_context 'version_stub'
+
   it 'uses the custom resource' do
     expect(chef_run).to create_bind_stub_zone('example.com')
     expect(chef_run).to create_bind_stub_zone('example.org')